+++ examples/imported/prometheus/prometheus/codeql-analysis/workflows/jobs.go
@@ -1,0 +1,11 @@
+package workflows
+
+import (
+	"github.com/lex00/wetwire-github-go/workflow"
+)
+
+var Analyze = workflow.Job{
+	Name:   "Analyze",
+	RunsOn: "ubuntu-latest",
+	Steps:  AnalyzeSteps,
+}